backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
coverage==7.10.7
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
librt==0.7.7
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
openai==2.15.0
packaging==25.0
pathspec==1.0.3
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
ruff==0.14.11
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tomli==2.4.0
tqdm==4.67.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
